+++ src/main/java/com/elmakers/mine/bukkit/protection/GriefPreventionAPI.java
@@ -2,6 +2,7 @@ package com.elmakers.mine.bukkit.protection;
 
 import me.ryanhamshire.GriefPrevention.Claim;
 import me.ryanhamshire.GriefPrevention.GriefPrevention;
+import org.bukkit.Location;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
@@ -24,6 +25,24 @@ public class GriefPreventionAPI
             if (claim != null) {
                 return claim.allowEdit(player) == null;
             }
+        }
+        return true;
+    }
+
+    public boolean hasBreakPermission(Player player, Block block) {
+        if (block != null && griefPrevention != null) {
+            Claim claim = griefPrevention.dataStore.getClaimAt(block.getLocation(), false, null);
+            if (claim != null) {
+                return claim.allowEdit(player) == null;
+            }
+        }
+        return true;
+    }
+
+    public boolean isPVPAllowed(Location location) {
+        if (location != null && griefPrevention != null && griefPrevention.config_pvp_noCombatInPlayerLandClaims) {
+            Claim claim = griefPrevention.dataStore.getClaimAt(location, false, null);
+            return (claim != null);
         }
         return true;
     }